Update to 1.0.4b
Added every skin in the game. Added every case in the game. Price of the skins were changed from "1000 credits" to "$100.0 credits". Prices are now more realistic, being as close as they can to the steam marketplace for most of the items. Chances in the cases have been fixed and auto-generated based on the price. Fixed a bug when buying or opening a skin where it wouldn't print in the chat the skin(for rare items). Fixed few run-time errors.